Move back tests for Study.tell to test_study.py
